requirements.txt
# Content Summarizer Dependencies

# YouTube 추출
youtube-transcript-api>=0.6.0
yt-dlp>=2024.10.22

# 로컬 비디오 추출
openai-whisper>=20231117
opencv-python>=4.8.0

# PDF 추출
pdfplumber>=0.10.0
pymupdf>=1.23.0
pytesseract>=0.3.10

# 웹 추출
trafilatura>=1.8.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# 유틸리티
python-dotenv>=1.0.0
rich>=13.0.0

# AI 노트 생성 (선택적)
anthropic>=0.18.0